// src/structures/enums.ts
export enum ESide {
  front = 0,
  top = 1,
  right = 2,
  back = 3,
  bottom = 4,
  left = 5,
}

export enum EColor {
  red = 0,
  green = 1,
  white = 2,
  yellow = 3,
  blue = 4,
  cyan = 5,
}

export enum ERotateDirection {
  rightBackward = -3,
  frontBackward = -2,
  topBackward = -1,
  frontForward = 1,
  topForward = 2,
  rightForward = 3,
}

export enum EEdgeRotation {
  forward = 1,
  backward = -1,
}